+++ src/serviceDefinition/apply.ts
@@ -1,6 +1,5 @@
-/*
- * Copyright 2022 steadybit GmbH. All rights reserved.
- */
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
 import { Response } from 'node-fetch';
 import colors from 'colors/safe';
@@ -9,7 +8,7 @@ import yaml from 'js-yaml';
 import { loadServiceDefinition, writeServiceDefinition } from './files';
 import { abortExecutionWithError } from '../errors';
 import { ServiceDefinition } from './types';
-import { executeApiCall } from '../api';
+import { executeApiCall } from '../api/http';
 import { v4 as uuidv4 } from 'uuid';
 import { addVcsTags } from '../vcs';