+++ src/serviceDefinition/delete.ts
@@ -1,10 +1,9 @@
-/*
- * Copyright 2022 steadybit GmbH. All rights reserved.
- */
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
 import { loadServiceDefinition } from './files';
 import { abortExecution } from '../errors';
-import { executeApiCall } from '../api';
+import { executeApiCall } from '../api/http';
 
 export interface Options {
   id?: string;