+++ src/serviceDefinition/show.test.ts
@@ -1,13 +1,12 @@
-/*
- * Copyright 2022 steadybit GmbH. All rights reserved.
- */
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
 import { loadServiceDefinition } from './files';
-import { executeApiCall } from '../api';
+import { executeApiCall } from '../api/http';
 import { executeShow } from './show';
 
 jest.mock('./files');
-jest.mock('../api');
+jest.mock('../api/http');
 
 const loadServiceDefinitionMock = loadServiceDefinition as jest.Mock;
 const executeApiCallMock = executeApiCall as jest.Mock;