+++ src/serviceDefinition/verify.ts
@@ -1,31 +1,10 @@
-/*
- * Copyright 2022 steadybit GmbH. All rights reserved.
- */
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
-import { blue, bold, gray, green, red } from 'colors/safe';
-
-import { ServiceState, Task, TaskState } from './types';
-import { abortExecutionWithError } from '../errors';
+import { getTaskCountByType, printTaskList, printTaskStateByType } from './taskList';
 import { loadServiceDefinition } from './files';
-import { executeApiCall } from '../api';
-
-const taskSuffix: Record<TaskState, string> = {
-  SUCCESS: 'ok',
-  FAILURE: 'failure',
-  PENDING: 'not checked',
-};
-
-const taskColor: Record<TaskState, typeof red> = {
-  SUCCESS: green,
-  FAILURE: red,
-  PENDING: blue,
-};
-
-const taskOrder: Record<TaskState, number> = {
-  SUCCESS: 0,
-  PENDING: 1,
-  FAILURE: 2,
-};
+import { bold, gray } from 'colors/safe';
+import { getState } from './api';
 
 interface Options {
   file: string;
@@ -33,105 +12,34 @@ interface Options {
   printMatrixContext?: boolean;
 }
 
-const openHelp = gray(`
+const openHelp = gray(
+  `
 Do you need more information? You can see more details within Steadybit's
-user interface. Try execute the following to open it:
+user interface. Try executing the following to open it:
 
                    ${bold('steadybit service open')}
-`.trim());
+`.trim()
+);
 
 export async function verify(options: Options) {
   const serviceDefinition = await loadServiceDefinition(options.file);
+  const state = await getState(serviceDefinition);
 
-  try {
-    const response = await executeApiCall({
-      method: 'GET',
-      path: `/api/service-states/${serviceDefinition.id}`,
-    });
-    const state = (await response.json()) as ServiceState;
-
-    printTaskList(options, state);
-
-    console.log();
-    console.log(openHelp);
-
-    const countByType = getTaskCountByType(state);
-    if (countByType.FAILURE > 0 || countByType.PENDING > 0) {
-      process.exit(1);
-    }
-  } catch (e) {
-    const error = await abortExecutionWithError(e, 'Failed to read state for service %s', serviceDefinition.id);
-    throw error;
-  }
-}
-
-function getTaskCountByType(state: ServiceState): Record<TaskState, number> {
-  const countByType: Record<TaskState, number> = {
-    SUCCESS: 0,
-    FAILURE: 0,
-    PENDING: 0,
-  };
-
-  for (const task of state.tasks) {
-    countByType[task.state]++;
-  }
-
-  return countByType;
-}
-
-function printTaskList(options: Options, state: ServiceState) {
-  console.log(bold('Tasks'));
-  console.log(bold('====='));
-
-  const sortedTasks = state.tasks.slice().sort((a, b) => {
-    let result = taskOrder[a.state] - taskOrder[b.state];
-    if (result === 0) {
-      result = a.definition.name.localeCompare(b.definition.name);
-    }
-    return result;
-  });
-
-  const countByType = getTaskCountByType(state);
-
-  const taskCountByCoordinate: Record<string, number> = {};
-  for (const task of sortedTasks) {
-    const key = getCoordinateKey(task);
-    taskCountByCoordinate[key] = (taskCountByCoordinate[key] ?? -1) + 1;
-  }
-
-  for (const task of sortedTasks) {
-    const key = getCoordinateKey(task);
-
-    console.log(
-      taskColor[task.state](' - %s (%s)'),
-      key,
-      taskSuffix[task.state]
-    );
-
-    if (options.printParameters) {
-      printJson('Parameters', task.parameters);
-    }
-
-    if (options.printMatrixContext || taskCountByCoordinate[key] > 1) {
-      printJson('Matrix Context', task.matrixContext);
-    }
-  }
+  printTaskList({
+    ...options,
+    colorsByTaskState: true,
+    renderTaskState: true
+  }, state.tasks);
 
   console.log();
-  console.log(taskColor.SUCCESS('Ok:          %d'), countByType.SUCCESS);
-  console.log(taskColor.PENDING('Not Checked: %d'), countByType.PENDING);
-  console.log(taskColor.FAILURE('Failure:     %d'), countByType.FAILURE);
-}
 
-function getCoordinateKey(task: Task): string {
-  return `${task.definition.name}@${task.definition.version}`;
-}
+  const countByType = getTaskCountByType(state.tasks);
+  printTaskStateByType(countByType);
 
-function printJson(title: string, obj: unknown): void {
-  const content = `${title}:\n${JSON.stringify(obj, undefined, 2)}\n`
-    .split('\n')
-    .map(s => '   ' + s)
-    .join('\n');
+  console.log();
+  console.log(openHelp);
 
-  console.log(gray(content));
+  if (countByType.FAILURE > 0 || countByType.PENDING > 0) {
+    process.exit(1);
+  }
 }